requirements.txt
# Production dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx[http2]==0.25.2
feedparser==6.0.10
pydantic-settings==2.1.0
python-dotenv==1.0.0
sqlmodel==0.0.14
aiosqlite==0.19.0
apscheduler==3.10.4
backoff==2.2.1
pyyaml==6.0.1
jinja2==3.1.2
python-multipart==0.0.6
chardet==5.2.0
# Auth dependencies
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
